requirements.txt
filelock==3.20.3
fsspec==2026.2.0
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
pillow==12.1.0
pypdf==6.6.2
pypdfium2==5.3.0
setuptools==80.10.2
sympy==1.14.0
torch==2.10.0
tqdm==4.67.3
typing-extensions==4.15.0
